Fix avoid_using_api lint tests on Linux
Banning a source file in the same library is not supported, and it does not work on Linux even if forced with package:this_app
