+++ example/analysis_options.yaml
@@ -1,4 +1,4 @@
-include: package:solid_lints/analysis_options.yaml
+include: ../analysis_options.yaml
 
 analyzer:
   plugins:
