+++ lint_test/avoid_using_api/class_source_ban/analysis_options.yaml
@@ -8,8 +8,8 @@ custom_lint:
       severity: warning
       entries:
         - class_name: BannedCodeUsage
-          source: package:class_source_ban
-          reason: "BannedCodeUsage from package:class_source_ban is not allowed"
+          source: package:external_source
+          reason: "BannedCodeUsage from package:external_source is not allowed"
         - class_name: UnmodifiableListView
           source: dart:collection
           reason: "UnmodifiableListView from dart:collection is not allowed"
