+++ lint_test/avoid_using_api/external_source/lib/banned_library.dart
