+++ lint_test/avoid_using_api/external_source/lib/external_source.dart
@@ -1,0 +1,20 @@
+class BannedCodeUsage {
+  BannedCodeUsage();
+  static String test2() {
+    return 'Hello World';
+  }
+
+  final String test4 = 'Hello World';
+
+  void test() {}
+
+  factory BannedCodeUsage.test3() {
+    return BannedCodeUsage();
+  }
+}
+
+const test2 = 'Hello World';
+
+void test() {}
+
+int banned = 5;
