+++ lint_test/avoid_using_api/external_source/pubspec.yaml
@@ -1,14 +1,10 @@
-name: identifier_class_source_ban
-description: A sample command-line application.
+name: external_source
+description: A starting point for Dart libraries or applications.
 version: 1.0.0
 publish_to: none
 
 environment:
   sdk: ^3.1.3
-
-dependencies:
-  external_source:
-    path: ../external_source
 
 dev_dependencies:
   lints: ^3.0.0
