+++ lint_test/avoid_using_api/identifier_class_source_ban/analysis_options.yaml
@@ -9,20 +9,20 @@ custom_lint:
       entries:
         - identifier: test4
           class_name: BannedCodeUsage
-          source: package:identifier_class_source_ban
-          reason: "BannedCodeUsage.test4 from the identifier_class_source_ban package is not allowed"
+          source: package:external_source
+          reason: "BannedCodeUsage.test4 from package:external_source is not allowed"
         - identifier: test3
           class_name: BannedCodeUsage
-          source: package:identifier_class_source_ban
-          reason: "BannedCodeUsage.test3 from the identifier_class_source_ban package is not allowed"
+          source: package:external_source
+          reason: "BannedCodeUsage.test3 from package:external_source is not allowed"
         - identifier: test2
           class_name: BannedCodeUsage
-          source: package:identifier_class_source_ban
-          reason: "BannedCodeUsage.test2 from the identifier_class_source_ban package is not allowed"
+          source: package:external_source
+          reason: "BannedCodeUsage.test2 from package:external_source is not allowed"
         - identifier: test
           class_name: BannedCodeUsage
-          source: package:identifier_class_source_ban
-          reason: "BannedCodeUsage.test from the identifier_class_source_ban package is not allowed"
+          source: package:external_source
+          reason: "BannedCodeUsage.test from package:external_source is not allowed"
         - class_name: UnmodifiableListView
           source: dart:collection
           reason: "UnmodifiableListView from dart:collection is not allowed"
