+++ lint_test/avoid_using_api/source_ban/lib/source_ban_test.dart
@@ -1,4 +1,4 @@
-import 'src/banned_library.dart';
+import 'package:external_source/banned_library.dart';
 
 void testingBannedCodeLint() {
   // expect_lint: avoid_using_api
