--- lint_test/avoid_using_api/source_excludes_ban/lib/src/banned_library.dart
@@ -1,1 +1,0 @@
-int banned = 5;
