+++ lint_test/avoid_using_api/source_includes_ban/pubspec.yaml
@@ -6,6 +6,10 @@ publish_to: none
 environment:
   sdk: ^3.1.3
 
+dependencies:
+  external_source:
+    path: ../external_source
+
 dev_dependencies:
   lints: ^3.0.0
   test: ^1.21.0
